+++ binance/spot/__init__.py
@@ -19,6 +19,7 @@ class Spot(API):
     from binance.spot._market import ui_klines
     from binance.spot._market import avg_price
     from binance.spot._market import ticker_24hr
+    from binance.spot._market import trading_day_ticker
     from binance.spot._market import ticker_price
     from binance.spot._market import book_ticker
     from binance.spot._market import rolling_window_ticker
@@ -41,6 +42,8 @@ class Spot(API):
     from binance.spot._trade import my_trades
     from binance.spot._trade import get_order_rate_limit
     from binance.spot._trade import query_prevented_matches
+    from binance.spot._trade import query_allocations
+    from binance.spot._trade import query_commission_rates
 
     # STREAMS
     from binance.spot._data_stream import new_listen_key
@@ -95,15 +98,9 @@ class Spot(API):
     from binance.spot._margin import get_small_liability_exchange_history
     from binance.spot._margin import get_a_future_hourly_interest_rate
     from binance.spot._margin import adjust_cross_margin_max_leverage
-
-    # Staking
-    from binance.spot._staking import staking_product_list
-    from binance.spot._staking import staking_purchase_product
-    from binance.spot._staking import staking_redeem_product
-    from binance.spot._staking import staking_product_position
-    from binance.spot._staking import staking_history
-    from binance.spot._staking import staking_set_auto_staking
-    from binance.spot._staking import staking_product_quota
+    from binance.spot._margin import margin_available_inventory
+    from binance.spot._margin import margin_manual_liquidation
+    from binance.spot._margin import liability_coin_leverage_bracket
 
     # WALLET
     from binance.spot._wallet import system_status
@@ -130,11 +127,13 @@ class Spot(API):
     from binance.spot._wallet import bnb_convertible_assets
     from binance.spot._wallet import convertible_coins
     from binance.spot._wallet import toggle_auto_convertion
+    from binance.spot._wallet import list_deposit_address
     from binance.spot._wallet import cloud_mining_trans_history
     from binance.spot._wallet import convert_transfer
     from binance.spot._wallet import convert_history
     from binance.spot._wallet import one_click_arrival_deposit_apply
     from binance.spot._wallet import balance
+    from binance.spot._wallet import delist_schedule_symbols
 
     # MINING
     from binance.spot._mining import mining_algo_list
@@ -212,22 +211,6 @@ class Spot(API):
     from binance.spot._blvt import redemption_record
     from binance.spot._blvt import user_limit_info
 
-    # BSwap
-    from binance.spot._bswap import bswap_pools
-    from binance.spot._bswap import bswap_liquidity
-    from binance.spot._bswap import bswap_liquidity_add
-    from binance.spot._bswap import bswap_liquidity_remove
-    from binance.spot._bswap import bswap_liquidity_operation_record
-    from binance.spot._bswap import bswap_request_quote
-    from binance.spot._bswap import bswap_swap
-    from binance.spot._bswap import bswap_swap_history
-    from binance.spot._bswap import bswap_pool_configure
-    from binance.spot._bswap import bswap_add_liquidity_preview
-    from binance.spot._bswap import bswap_remove_liquidity_preview
-    from binance.spot._bswap import bswap_unclaimed_rewards
-    from binance.spot._bswap import bswap_claim_rewards
-    from binance.spot._bswap import bswap_claimed_rewards
-
     # FIAT
     from binance.spot._fiat import fiat_order_history
     from binance.spot._fiat import fiat_payment_history
@@ -252,15 +235,6 @@ class Spot(API):
     from binance.spot._crypto_loan import loan_collateral_data
     from binance.spot._crypto_loan import loan_collateral_rate
     from binance.spot._crypto_loan import loan_customize_margin_call
-    from binance.spot._crypto_loan import flexible_loan_adjust_ltv
-    from binance.spot._crypto_loan import flexible_loan_assets_data
-    from binance.spot._crypto_loan import flexible_loan_borrow_history
-    from binance.spot._crypto_loan import flexible_loan_borrow
-    from binance.spot._crypto_loan import flexible_loan_collateral_assets_data
-    from binance.spot._crypto_loan import flexible_loan_ltv_adjustment_history
-    from binance.spot._crypto_loan import flexible_loan_ongoing_orders
-    from binance.spot._crypto_loan import flexible_loan_repay
-    from binance.spot._crypto_loan import flexible_loan_repayment_history
 
     # PAY
     from binance.spot._pay import pay_history
@@ -271,6 +245,9 @@ class Spot(API):
     from binance.spot._convert import send_quote_request
     from binance.spot._convert import accept_quote
     from binance.spot._convert import order_status
+    from binance.spot._convert import place_limit_order
+    from binance.spot._convert import cancel_limit_order
+    from binance.spot._convert import query_limit_open_order
     from binance.spot._convert import get_convert_trade_history
 
     # REBATE