+++ tests/spot/trade/test_query_commission_rates.py
@@ -1,0 +1,30 @@
+import responses
+
+from binance.spot import Spot as Client
+from tests.util import random_str
+from urllib.parse import urlencode
+from tests.util import mock_http_response
+
+
+mock_item = {"key_1": "value_1", "key_2": "value_2"}
+mock_exception = {"code": -1, "msg": "error message"}
+
+key = random_str()
+secret = random_str()
+
+send_params = {"symbol": "BTCUSDT", "recvWindow": 5000}
+expected_params = {"symbol": "BTCUSDT", "recvWindow": 5000}
+
+
+@mock_http_response(
+    responses.GET,
+    "/api/v3/account/commission\\?" + urlencode(expected_params),
+    mock_item,
+    200,
+)
+def test_query_commission_rates():
+    """Tests the API endpoint to get current account commission rates"""
+
+    client = Client(key, secret)
+    response = client.query_commission_rates(**send_params)
+    response.should.equal(mock_item)